test(cketh): split the anvil client into its own test_utils module
Move the anvil node client and its ABI/solc helpers out of live_scan
into a dedicated ic_cketh_test_utils::anvil module, shared by both the
standalone batcher tests and the live-scan harness.

Also close a balance-scan race in credit_deposit: write every token
balance before placing any code, so the fail-loud batcher only advances
an address once all balances are in place — a concurrent scan can no
longer observe a partially-credited address (addresses the review
comment on candidates_found nondeterminism).
